requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pdfplumber==0.10.2
python-docx==0.8.11
spacy==3.6.1
sentence-transformers==2.7.0
langchain==0.0.300
streamlit==1.27.0
scikit-learn==1.3.0
xgboost==2.1.0
requests==2.31.0
plotly==5.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.0
transformers==4.36.2
torch==2.1.0
huggingface_hub==0.24.0